// ui/src/generated/types/device.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Mercury
 * Control and manages real Smartphone devices from browser and remercuryul apis
 * OpenAPI spec version: 1.4.7
 */
import type { DeviceBattery } from './deviceBattery'
import type { DeviceBrowser } from './deviceBrowser'
import type { DeviceCapabilities } from './deviceCapabilities'
import type { DeviceCpu } from './deviceCpu'
import type { DeviceDisplay } from './deviceDisplay'
import type { DeviceGroup } from './deviceGroup'
import type { DeviceMemory } from './deviceMemory'
import type { DeviceNetwork } from './deviceNetwork'
import type { DeviceOwner } from './deviceOwner'
import type { DevicePhone } from './devicePhone'
import type { DeviceProvider } from './deviceProvider'
import type { DeviceReverseForwardsItem } from './deviceReverseForwardsItem'
import type { DeviceService } from './deviceService'

export interface Device {
  _id?: string
  abi?: string
  adbPort?: number
  airplaneMode?: boolean
  battery?: DeviceBattery
  bookedBefore?: number
  browser?: DeviceBrowser
  capabilities?: DeviceCapabilities
  channel?: string
  connectPort?: number
  cpu?: DeviceCpu
  cpuPlatform?: string
  createdAt?: string
  deviceType?: string
  display?: DeviceDisplay
  group?: DeviceGroup
  image?: string
  iosClientChannel?: string
  likelyLeaveReason?: string
  logs_enabled?: boolean
  macAddress?: string
  manufacturer?: string
  marketName?: string
  memory?: DeviceMemory
  model?: string
  name?: string
  network?: DeviceNetwork
  notes?: string
  openGLESVersion?: string
  /** @nullable */
  operator?: string | null
  /** @nullable */
  owner?: DeviceOwner
  phone?: DevicePhone
  place?: string
  platform?: string
  presenceChangedAt?: string
  present?: boolean
  product?: string
  provider?: DeviceProvider
  ram?: string
  ready?: boolean
  releasedAt?: string
  remoteConnect?: boolean
  /** @nullable */
  remoteConnectUrl?: string | null
  reverseForwards?: DeviceReverseForwardsItem[]
  screenPort?: number
  sdk?: string
  serial?: string
  service?: DeviceService
  status?: number
  statusChangedAt?: string
  storageId?: string
  /** @nullable */
  usage?: string | null
  usageChangedAt?: string
  using?: boolean
  version?: string
}