// ui/src/generated/types/userResponse.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Mercury
 * Control and manages real Smartphone devices from browser and remercuryul apis
 * OpenAPI spec version: 1.4.7
 */
import type { User } from './user'

export interface UserResponse {
  description: string
  success: boolean
  user?: User
}