// ui/src/generated/types/getUsersParams.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Mercury
 * Control and manages real Smartphone devices from browser and remercuryul apis
 * OpenAPI spec version: 1.4.7
 */

export type GetUsersParams = {
  /**
   * Comma-separated list of user fields; only listed fields will be returned in response
   */
  fields?: string
}