// ui/src/generated/types/getGroupsParams.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Mercury
 * Control and manages real Smartphone devices from browser and remercuryul apis
 * OpenAPI spec version: 1.4.7
 */

export type GetGroupsParams = {
  /**
   * Comma-seperated list of fields; only listed fields will be returned in response
   */
  fields?: string
  /**
   * Selects the groups for which you are the owner (true) or a simple member (false); note that by not providing this parameter, it means all groups to which you belong are selected
   */
  owner?: boolean
}